test: move scale tests to e2e because slow
