+++ test/cli-ux/styled/table.test.ts
@@ -318,36 +318,5 @@ describe('styled/table', () => {
  123 supertable-test-1${ws}\n`)
     })
   })
-
-  describe('scale tests', () => {
-    const bigRows = 150_000
-    fancy
-    .stdout()
-    .timeout(600_000)
-    .end('very tall tables don\'t exceed stack depth', output => {
-      const data = Array.from({length: bigRows}).fill({id: '123', name: 'foo', value: 'bar'}) as Record<string, unknown>[]
-      const tallColumns = {
-        id: {header: 'ID'},
-        name: {},
-        value: {header: 'TEST'},
-      }
-
-      CliUx.ux.table(data, tallColumns)
-      expect(output.stdout).to.include('ID')
-    })
-
-    fancy
-    .stdout()
-    .timeout(600_000)
-    .end('very tall, wide tables don\'t exceed stack depth', output => {
-      const columns = 100
-      const row = Object.fromEntries(Array.from({length: columns}).map((_, i) => [`col${i}`, 'foo']))
-      const data = Array.from({length: bigRows}).fill(row) as Record<string, unknown>[]
-      const bigColumns = Object.fromEntries(Array.from({length: columns}).map((_, i) => [`col${i}`, {header: `col${i}`.toUpperCase()}]))
-
-      CliUx.ux.table(data, bigColumns)
-      expect(output.stdout).to.include('COL1')
-    })
-  })
 })
